# Resound-UI/Builtiful-Interface/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  description: Resound customer signal intelligence API
  version: 0.1.0
paths:
...
  /brands:
    get:
      tags:
      - brands
      summary: List Brands
      operationId: listBrands
      parameters:
      - name: X-Resound-Organization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Organization
      - name: X-Resound-Team
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Team
      - name: X-Resound-User
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Brand'
                title: Response Listbrands
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /brands/{brandId}/stats/{period}:
    get:
      tags:
      - brands
      summary: Get Brand Stats
      operationId: getBrandStats
      parameters:
      - name: brandId
        in: path
        required: true
        schema:
          type: string
          title: Brandid
      - name: period
        in: path
        required: true
        schema:
          enum:
          - 24h
          - 7d
          - 30d
          - qtd
          type: string
          title: Period
      - name: X-Resound-Organization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Organization
      - name: X-Resound-Team
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Team
      - name: X-Resound-User
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandStats'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /routes:
    get:
      tags:
      - routes
      - feedback
      summary: List Routes
      operationId: listRoutes
      parameters:
      - name: brandId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Brandid
      - name: period
        in: query
        required: false
        schema:
          enum:
          - 24h
          - 7d
          - 30d
          - qtd
          type: string
          default: 7d
          title: Period
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: X-Resound-Organization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Organization
      - name: X-Resound-Team
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Team
      - name: X-Resound-User
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RouteAudit'
                title: Response Listroutes
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /routes/{routeId}/reroute:
    patch:
      tags:
      - routes
      - feedback
      summary: Reroute Signal
      operationId: rerouteSignal
      parameters:
      - name: routeId
        in: path
        required: true
        schema:
          type: integer
          title: Routeid
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      - name: X-Resound-Organization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Organization
      - name: X-Resound-Team
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Team
      - name: X-Resound-User
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RerouteInput'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteAudit'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /routes/{routeId}/feedback:
    post:
      tags:
      - routes
      - feedback
      summary: Submit Feedback
      operationId: submitFeedback
      parameters:
      - name: routeId
        in: path
        required: true
        schema:
          type: integer
          title: Routeid
      - name: X-Resound-Organization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Organization
      - name: X-Resound-Team
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Team
      - name: X-Resound-User
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackInput'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackEvent'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /telemetry/llm:
    get:
      tags:
      - operations
      summary: Get Llm Telemetry
      operationId: getLlmTelemetry
      parameters:
      - name: brandId
        in: query
        required: true
        schema:
          type: string
          title: Brandid
      - name: period
        in: query
        required: false
        schema:
          enum:
          - 24h
          - 7d
          - 30d
          - qtd
          type: string
          default: 7d
          title: Period
      - name: X-Resound-Organization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Organization
      - name: X-Resound-Team
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Team
      - name: X-Resound-User
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMTelemetry'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /evaluations/summary:
    get:
      tags:
      - operations
      summary: Get Evaluation Summary
      operationId: getEvaluationSummary
      parameters:
      - name: brandId
        in: query
        required: true
        schema:
          type: string
          title: Brandid
      - name: period
        in: query
        required: false
        schema:
          enum:
          - 24h
          - 7d
          - 30d
          - qtd
          type: string
          default: 7d
          title: Period
      - name: X-Resound-Organization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Organization
      - name: X-Resound-Team
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Team
      - name: X-Resound-User
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Brand:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        slug:
          type: string
          title: Slug
        description:
          type: string
          title: Description
        primaryContact:
          type: string
          title: Primarycontact
        sourcesActive:
          items:
            type: string
          type: array
          title: Sourcesactive
        lastIngested:
          anyOf:
          - type: string
          - type: 'null'
          title: Lastingested
        tagline:
          type: string
          title: Tagline
        ownerOptions:
          items:
            $ref: '#/components/schemas/OwnerOption'
          type: array
          title: Owneroptions
      type: object
      required:
      - id
      - name
      - slug
      - description
      - primaryContact
      - sourcesActive
      - lastIngested
      - tagline
      title: Brand
    BrandStats:
      properties:
        brandId:
          type: string
          title: Brandid
        period:
          type: string
          title: Period
        netSentiment:
          type: integer
          title: Netsentiment
        netSentimentDelta:
          type: integer
          title: Netsentimentdelta
        criticalCount:
          type: integer
          title: Criticalcount
        criticalDelta:
          type: integer
          title: Criticaldelta
        totalVolume:
          type: integer
          title: Totalvolume
        volumeDelta:
          type: number
          title: Volumedelta
        sourceMix:
          items:
            $ref: '#/components/schemas/SourceStat'
          type: array
          title: Sourcemix
        sentimentBreakdown:
          $ref: '#/components/schemas/SentimentBreakdown'
        topEmergingIssue:
          $ref: '#/components/schemas/PatternSummary'
        lastIngested:
          anyOf:
          - type: string
          - type: 'null'
          title: Lastingested
      type: object
      required:
      - brandId
      - period
      - netSentiment
      - netSentimentDelta
      - criticalCount
      - criticalDelta
      - totalVolume
      - volumeDelta
      - sourceMix
      - sentimentBreakdown
      - topEmergingIssue
      title: BrandStats
...
    EvaluationSummary:
      properties:
        brandId:
          type: string
          title: Brandid
        reportRunsByStatus:
          additionalProperties:
            type: integer
          type: object
          title: Reportrunsbystatus
        sourceFailureCount:
          type: integer
          title: Sourcefailurecount
        totalLlmCostUsd:
          type: number
          title: Totalllmcostusd
      type: object
      required:
      - brandId
      - reportRunsByStatus
      - sourceFailureCount
      - totalLlmCostUsd
      title: EvaluationSummary
    FeedbackEvent:
      properties:
        id:
          type: integer
          title: Id
        routeId:
          type: integer
          title: Routeid
        correct:
          type: boolean
          title: Correct
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        createdAt:
          type: string
          title: Createdat
      type: object
      required:
      - id
      - routeId
      - correct
      - createdAt
      title: FeedbackEvent
    FeedbackInput:
      properties:
        correct:
          type: boolean
          title: Correct
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        actioned:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Actioned
        submittedBy:
          anyOf:
          - type: string
          - type: 'null'
          title: Submittedby
      type: object
      required:
      - correct
      title: FeedbackInput
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LLMTelemetry:
      properties:
        brandId:
          type: string
          title: Brandid
        period:
          type: string
          enum:
          - 24h
          - 7d
          - 30d
          - qtd
          title: Period
        costs:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Costs
        latency:
          additionalProperties:
            additionalProperties:
              type: number
            type: object
          type: object
          title: Latency
        fallbackRate:
          additionalProperties:
            additionalProperties:
              type: number
            type: object
          type: object
          title: Fallbackrate
      type: object
      required:
      - brandId
      - period
      - costs
      - latency
      - fallbackRate
      title: LLMTelemetry
...
    OwnerOption:
      properties:
        owner:
          type: string
          title: Owner
        label:
          type: string
          title: Label
        hint:
          type: string
          title: Hint
      type: object
      required:
      - owner
      - label
      - hint
      title: OwnerOption
...
    PatternSummary:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        area:
          type: string
          title: Area
        signalCount:
          type: integer
          title: Signalcount
        weeklyVelocity:
          type: integer
          title: Weeklyvelocity
        velocityMultiple:
          type: number
          title: Velocitymultiple
      type: object
      required:
      - id
      - name
      - area
      - signalCount
      - weeklyVelocity
      - velocityMultiple
      title: PatternSummary
...
    RerouteInput:
      properties:
        owner:
          type: string
          title: Owner
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        submittedBy:
          anyOf:
          - type: string
          - type: 'null'
          title: Submittedby
      type: object
      required:
      - owner
      title: RerouteInput
...
    RouteAudit:
      properties:
        id:
          type: integer
          title: Id
        signalId:
          type: integer
          title: Signalid
        owner:
          type: string
          title: Owner
        area:
          type: string
          title: Area
        severity:
          type: string
          title: Severity
        sentiment:
          type: string
          title: Sentiment
        source:
          type: string
          title: Source
        content:
          type: string
          title: Content
        summary:
          type: string
          title: Summary
        confidence:
          type: number
          title: Confidence
        ruleMatched:
          anyOf:
          - type: string
          - type: 'null'
          title: Rulematched
        reroutedFrom:
          anyOf:
          - type: string
          - type: 'null'
          title: Reroutedfrom
        createdAt:
          type: string
          title: Createdat
        feedbackCorrect:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Feedbackcorrect
      type: object
      required:
      - id
      - signalId
      - owner
      - area
      - severity
      - sentiment
      - source
      - content
      - summary
      - confidence
      - ruleMatched
      - createdAt
      - feedbackCorrect
      title: RouteAudit
    SentimentBreakdown:
      properties:
        positive:
          type: number
          title: Positive
        neutral:
          type: number
          title: Neutral
        negative:
          type: number
          title: Negative
      type: object
      required:
      - positive
      - neutral
      - negative
      title: SentimentBreakdown
...
    SourceStat:
      properties:
        source:
          type: string
          title: Source
        count:
          type: integer
          title: Count
        pct:
          type: number
          title: Pct
      type: object
      required:
      - source
      - count
      - pct
      title: SourceStat
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
servers:
- url: /api
  description: Base API path